requirements.txt
numpy==2.4.2
opencv-python==4.13.0.90
pdf2image==1.17.0
pillow==12.1.0
